pyproject.toml
[project]
name = "zapsystem"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.7",
    "python-decouple==3.8",
]
